chore: move tooltip component
