+++ apps/web/src/components/Composer/Actions/AccessSettings/index.tsx
@@ -1,4 +1,3 @@
-import { Tooltip } from '@components/UI/Tooltip';
 import { LockClosedIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
 import clsx from 'clsx';
@@ -6,7 +5,7 @@ import { motion } from 'framer-motion';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { useAccessSettingsStore } from 'src/store/access-settings';
-import { HelpTooltip, Modal } from 'ui';
+import { HelpTooltip, Modal, Tooltip } from 'ui';
 
 import BasicSettings from './BasicSettings';
