+++ apps/web/src/components/Composer/Actions/CollectSettings/index.tsx
@@ -1,4 +1,3 @@
-import { Tooltip } from '@components/UI/Tooltip';
 import GetModuleIcon from '@components/utils/GetModuleIcon';
 import { CashIcon } from '@heroicons/react/outline';
 import { getModule } from '@lib/getModule';
@@ -7,7 +6,7 @@ import { motion } from 'framer-motion';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { useCollectModuleStore } from 'src/store/collect-module';
-import { Modal } from 'ui';
+import { Modal, Tooltip } from 'ui';
 
 import CollectForm from './CollectForm';
