+++ apps/web/src/components/Composer/Actions/Giphy/index.tsx
@@ -1,5 +1,4 @@
 import Loader from '@components/Shared/Loader';
-import { Tooltip } from '@components/UI/Tooltip';
 import type { IGif } from '@giphy/js-types';
 import { PhotographIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
@@ -8,7 +7,7 @@ import dynamic from 'next/dynamic';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { usePublicationStore } from 'src/store/publication';
-import { Modal } from 'ui';
+import { Modal, Tooltip } from 'ui';
 
 const GifSelector = dynamic(() => import('./GifSelector'), {
   loading: () => <Loader message={t`Loading GIFs`} />
