+++ apps/web/src/components/Composer/Post/New.tsx
@@ -1,4 +1,3 @@
-import { Tooltip } from '@components/UI/Tooltip';
 import { PencilAltIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
 import { useRouter } from 'next/router';
@@ -6,7 +5,7 @@ import type { FC, ReactNode } from 'react';
 import { useEffect, useState } from 'react';
 import { useAppStore } from 'src/store/app';
 import { usePublicationStore } from 'src/store/publication';
-import { Card, Image, Modal } from 'ui';
+import { Card, Image, Modal, Tooltip } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import getAvatar from 'utils/getAvatar';
