+++ apps/web/src/components/Profile/Badges/Worldcoin.tsx
@@ -1,9 +1,9 @@
-import { Tooltip } from '@components/UI/Tooltip';
 import { CheckCircleIcon } from '@heroicons/react/solid';
 import { Trans } from '@lingui/macro';
 import { STATIC_IMAGES_URL } from 'data/constants';
 import type { Profile } from 'lens';
 import type { FC } from 'react';
+import { Tooltip } from 'ui';
 
 interface WorldcoinProps {
   profile: Profile;
