+++ apps/web/src/components/Profile/Filters/MediaFilter.tsx
@@ -1,11 +1,11 @@
 import MenuTransition from '@components/Shared/MenuTransition';
-import { Tooltip } from '@components/UI/Tooltip';
 import { Menu } from '@headlessui/react';
 import { AdjustmentsIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import clsx from 'clsx';
 import type { ChangeEvent } from 'react';
 import { useProfileFeedStore } from 'src/store/profile-feed';
+import { Tooltip } from 'ui';
 import { Checkbox } from 'ui/Checkbox';
 
 const MediaFilter = () => {
