+++ apps/web/src/components/Publication/Actions/Analytics/index.tsx
@@ -1,6 +1,5 @@
 import Beta from '@components/Shared/Badges/Beta';
 import Loader from '@components/Shared/Loader';
-import { Tooltip } from '@components/UI/Tooltip';
 import useStaffMode from '@components/utils/hooks/useStaffMode';
 import { ChartBarIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
@@ -12,7 +11,7 @@ import { useRouter } from 'next/router';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { useAppStore } from 'src/store/app';
-import { Modal } from 'ui';
+import { Modal, Tooltip } from 'ui';
 import isFeatureEnabled from 'utils/isFeatureEnabled';
 
 const Stats = dynamic(() => import('./Stats'), {
