+++ apps/web/src/components/Publication/Actions/Collect/index.tsx
@@ -1,5 +1,4 @@
 import Loader from '@components/Shared/Loader';
-import { Tooltip } from '@components/UI/Tooltip';
 import GetModuleIcon from '@components/utils/GetModuleIcon';
 import { CollectionIcon } from '@heroicons/react/outline';
 import { CollectionIcon as CollectionIconSolid } from '@heroicons/react/solid';
@@ -13,7 +12,7 @@ import dynamic from 'next/dynamic';
 import type { FC } from 'react';
 import { useEffect, useState } from 'react';
 import { PUBLICATION } from 'src/tracking';
-import { Modal } from 'ui';
+import { Modal, Tooltip } from 'ui';
 import humanize from 'utils/humanize';
 import nFormatter from 'utils/nFormatter';
