+++ apps/web/src/components/Publication/Actions/Comment.tsx
@@ -1,10 +1,10 @@
-import { Tooltip } from '@components/UI/Tooltip';
 import { ChatAlt2Icon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import { motion } from 'framer-motion';
 import type { Publication } from 'lens';
 import Link from 'next/link';
 import type { FC } from 'react';
+import { Tooltip } from 'ui';
 import humanize from 'utils/humanize';
 import nFormatter from 'utils/nFormatter';
