+++ apps/web/src/components/Publication/Actions/Like.tsx
@@ -1,5 +1,4 @@
 import type { ApolloCache } from '@apollo/client';
-import { Tooltip } from '@components/UI/Tooltip';
 import { HeartIcon, SunIcon } from '@heroicons/react/outline';
 import { HeartIcon as HeartIconSolid, SunIcon as SunIconSolid } from '@heroicons/react/solid';
 import { Mixpanel } from '@lib/mixpanel';
@@ -17,6 +16,7 @@ import toast from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { usePreferencesStore } from 'src/store/preferences';
 import { PUBLICATION } from 'src/tracking';
+import { Tooltip } from 'ui';
 import hasGm from 'utils/hasGm';
 import { publicationKeyFields } from 'utils/keyFields';
 import nFormatter from 'utils/nFormatter';
