+++ apps/web/src/components/Publication/Actions/index.tsx
@@ -1,4 +1,3 @@
-import { Tooltip } from '@components/UI/Tooltip';
 import useModMode from '@components/utils/hooks/useModMode';
 import { LockClosedIcon } from '@heroicons/react/solid';
 import { stopEventPropagation } from '@lib/stopEventPropagation';
@@ -6,6 +5,7 @@ import { t } from '@lingui/macro';
 import type { ElectedMirror, Publication } from 'lens';
 import type { FC } from 'react';
 import { useAppStore } from 'src/store/app';
+import { Tooltip } from 'ui';
 
 import Analytics from './Analytics';
 import Collect from './Collect';
