+++ packages/ui/index.ts
@@ -15,3 +15,4 @@ export { Spinner } from './Spinner';
 export { default as TabButton } from './TabButton';
 export { TextArea } from './TextArea';
 export { Toggle } from './Toggle';
+export { Tooltip } from './Tooltip';
